Ship the crayio layer, and fix the integer width that shipping exposed
port/crayio/ becomes an installable compatibility layer rather than a
test fixture: -DLIBSPILL_WITH_CRAYIO=ON builds libspill_crayio, installs
its header, and exports libspill::spill_crayio.

This does not reopen 6a. That section withdrew crayio as an adoption
target and the reasons are unchanged -- Dalton and LSDalton are not
actively developed, MADNESS's copy sits in a full-CI module unchanged
since 1999, and NWChem's two copies have had no substantive change since
1995 while NWChem itself took 665 commits in two years. Easy to replace
and going to be replaced are different claims and only the first holds.
Shipping removes friction for anyone who wants it; it does not create a
maintainer who does, and none of the four should count toward criterion
1. What it buys is that a code can link a library instead of vendoring a
file out of somebody's test directory.

It is a separate library on purpose. wopen_, wclose_, getwa_ and putwa_
are global Fortran symbols, so putting them in libspill itself would
collide with any code that still has its own crayio.c, or silently
interpose one on the other. libspill.so exports none of them.

Shipping surfaced a defect the test fixture had hidden. The shim typed
the Fortran integer as long and claimed it "only ever widens it, so
either works". Every crayio argument arrives by pointer, so there is no
widening anywhere: reading a caller's int32 through an int64 pointer
reads four bytes it never wrote. A 32-bit-INTEGER caller got ierr = -1
from every call. VAR_INT64 is unset by default in every copy, so the
broken width was the default one.

Same defect and same cause as 6e found in the OpenMolcas shims: Fortran
externals carry no interface, so nothing in the toolchain objects. The
type now mirrors each copy's own #if defined(VAR_INT64), LIBSPILL_CRAYIO_I8
selects it, and the test is built and run at both widths.

+++ port/crayio/crayio_libspill.c
@@ -35,7 +35,7 @@
 #include <string.h>
 
 #include "libspill.h"
-#include "crayio_shim.h"
+#include "crayio_libspill.h"
 
 /* One store per unit; the whole word-addressed space is one record, because
  * crayio addresses a unit by offset alone and has no notion of a key. */
@@ -92,7 +92,7 @@ void FSYM(wopen)(const CRAY_INT *unit, const char *name, const CRAY_INT *lennam,
         memcpy(nm, name, k);
         nm[k] = '\0';
     } else {
-        sprintf(nm, "fort.%.2ld", u);                /* crayio's own default */
+        sprintf(nm, "fort.%.2d", (int)u);            /* crayio's own default */
     }
 
     ls_opts_default(&o);
